# .eh_frame: Remove expensive opcode dispatch