.eh_frame: Remove expensive opcode to function dispatch.

## Test Plan

```
$ go test -v github.com/parca-dev/parca-agent/pkg/stack/unwind -bench=. -count=20
```

```
$ benchstat /tmp/old /tmp/new
name                                  old time/op    new time/op    delta
ParsingLibcDwarfUnwindInformation-12    34.2ms ±28%    26.7ms ±33%  -21.86%  (p=0.000 n=20+20)

name                                  old alloc/op   new alloc/op   delta
ParsingLibcDwarfUnwindInformation-12    32.7MB ± 0%    32.7MB ± 0%     ~     (p=0.473 n=20+20)

name                                  old allocs/op  new allocs/op  delta
ParsingLibcDwarfUnwindInformation-12      150k ± 0%      150k ± 0%     ~     (p=1.000 n=20+20)
```